updated battle_mode, game_tools, normal_mode, start_game, main, requirment
