+++ MoveNow/.gitignore
@@ -7,6 +7,7 @@ players/target_posedata/image
 players/target_posedata/important
 players/target_posedata/old
 players/target_posedata/*.jpg
+spare.py
 #                                     __               __  __                   _
 #   ____ ____  ____  ___  _________ _/ /  ____  __  __/ /_/ /_  ____  ____     (_)___ _____  ____  ________  _____
 #  / __ `/ _ \/ __ \/ _ \/ ___/ __ `/ /  / __ \/ / / / __/ __ \/ __ \/ __ \   / / __ `/ __ \/ __ \/ ___/ _ \/ ___/
